remove comments/unused code, rename components/files
